# feat: use slab allocation for strings and bytes