@nrwiersma 2023-05-28
It allows specific errors, which are helpful. Can be renamed to op perhaps